+++ src/helpers/mod.rs
@@ -1,6 +1,3 @@
-mod echo_x_request_id;
-pub use echo_x_request_id::echo_x_request_id;
-
 mod with_shared_data;
 pub use with_shared_data::with_shared_data;
 
@@ -12,3 +9,5 @@ pub use unit_conversion::convert_metres_to_degrees;
 
 mod mean_angle;
 pub use mean_angle::mean_angle;
+
+pub mod serde_helpers;