+++ src/routes/mod.rs
@@ -1,5 +1,8 @@
 mod geoprocessing;
 
+mod echo_x_request_id;
+pub use echo_x_request_id::echo_x_request_id;
+
 pub mod query_parameters;
 
 mod lines;
@@ -10,3 +13,9 @@ pub use points::points;
 
 mod lines_batch;
 pub use lines_batch::lines_batch;
+
+mod custom_rejection_handler;
+pub use custom_rejection_handler::custom_rejection_handler;
+
+mod load_data_and_combine_filters;
+pub use load_data_and_combine_filters::load_data_and_combine_filters;