+++ changelog.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [Unreleased] 2023-06-12
+
+- drop support for case insensitive query
+- slk_from and slk_to are now option. One or either can be omitted to get the
+  road "up to" or "starting from" its endpoints
+- Changed CLI `--addr` to `--ip-address` because abbreviations are annoying
+- reject requests with unknown fields (for example if `cwy=` is misspelled as `cway=`)
+- fixed `x-request-id` now echo's properly on rejection
+- invalid values of `cwy=` will now cause the request to be rejected
+
 ## [0.10.2] 2023-06-11
 
 - Refactoring to make main.rs short and easy to understand and decouple